Improve canvas sizing mode, fix preview A/R
This commit changes the canvas sizing mode; it is now going to either
fill available space, or have a fixed size, on both directions. This
makes it behave more intuitively, and allows us to properly respect
the canvas aspect ratio in the preview. This is reflected in the UI.

That change also includes a bunch of cleanup (including addressing
Lint complaints) of the project code.

It also includes a change to how points are drawn in the gradient, to
avoid the points being squashed for non-square canvases.

Lastly, it adds an explicit output image size in the export scale
selector, allowing a better understanding of what each option means.